--- delectable-client/src/app/component/recipe-form2/recipe-form2.component.html
@@ -1,76 +1,0 @@
-<form [formGroup]="recipeForm" (ngSubmit)="onSubmit()">
-   <!-- Directions -->
-   <div class="form-group form-row" formArrayName="directions">
-      <div class="col">
-         <h5>Directions</h5>
-      </div>
-      <table class="table table-sm table-borderless">
-         <thead>
-            <tr>
-               <th style="width: 5%">#</th>
-               <th style="width: 95%">Instruction</th>
-            </tr>
-         </thead>
-         <tbody>
-            <tr *ngFor="let direction of directions.controls; let i = index" formGroupName="{{i}}">
-               <td>{{direction.controls.step.value}}</td>
-               <td><input type="text" class="form-control" formControlName="instructions"></td>
-               <td><button class="btn btn-danger" type="button" (click)=removeDirection(direction)>-</button></td>
-            </tr>
-            <tr>
-               <td></td>
-               <td></td>
-               <td><button class="btn btn-success" type="button" (click)=addDirection()>+</button></td>
-            </tr>
-         </tbody>
-      </table>
-   </div>
-
-   <!-- Ingreidnets -->
-   <div class="form-group form-row" formArrayName="ingredients">
-      <div class="col">
-         <h5>Ingredients</h5>
-      </div>
-      <table class="table table-hover table-borderless table-sm">
-         <thead>
-            <tr>
-               <th style="width: 60%">Item</th>
-               <th style="width: 15%">Quantity</th>
-               <th style="width: 25%">Serving</th>
-            </tr>
-         </thead>
-         <tbody>
-            <tr *ngFor="let ingredient of ingredients.controls; let i = index" formGroupName="{{i}}">
-               <input type="hidden" class=form-control formControlName="recipe">
-               <td>
-                  <select class="form-control" formControlName="pantry">
-                     <option *ngFor="let pantry of pantryItemList" [ngValue]=pantry.id>{{pantry.name}}</option>
-                  </select>
-               </td>
-               <td><input type="text" class="form-control" formControlName="quantity"></td>
-               <td><input type="text" class="form-control" formControlName="servingType"></td>
-               <td><button class="btn btn-danger" type="button" (click)=removeIngredient(i)>-</button></td>
-            </tr>
-            <tr>
-               <td></td>
-               <td></td>
-               <td></td>
-               <td><button class="btn btn-success" type="button" (click)=addIngredient()>+</button></td>
-            </tr>
-         </tbody>
-      </table>
-   </div>
-   <input type="hidden" class="form-control" id="id" formControlName="id">
-   <input type="hidden" class="form-control" id="name" formControlName="name">
-   <input type="hidden" class="form-control" id="source" formControlName="source">
-   <input type="hidden" class="form-control" id="prepTime" formControlName="prepTime">
-   <input type="hidden" class="form-control" id="cookTime" formControlName="cookTime">
-   <div class="row">
-      <div class="col-auto">
-         <button class="btn btn-dark" type="submit">Submit</button>
-      </div>
-   </div>
-</form>
-<p>
-Form Value: {{ recipeForm.value | json }}
-</p>
