+++ delectable-client/src/app/delectable/pantry/component/pantry-detail/pantry-detail.component.html
