+++ delectable-client/src/app/delectable/pantry/component/pantry-list/pantry-list.component.html
