+++ delectable-client/src/app/delectable/pantry/component/pantry-list/pantry-list.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
-import { Pantry } from '../../model/pantry';
-import { PantryService } from '../../service/pantry.service';
+import { Pantry } from '../../pantry';
+import { PantryService } from '../../pantry.service';
 
 @Component({
   selector: 'app-pantry-list',
