+++ delectable-client/src/app/delectable/recipe/component/recipe-detail/recipe-detail.component.html
