+++ delectable-client/src/app/delectable/recipe/component/recipe-list/recipe-list.component.spec.ts
