+++ delectable-client/src/app/delectable/recipe/recipe.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { HttpClient, HttpHeaders } from '@angular/common/http';
-import { Recipe } from '../model/recipe';
+import { Recipe } from './recipe';
 import { Observable } from 'rxjs'; 
 
 @Injectable()
