+++ delectable-client/src/app/delectable/schedule/schedule-list/schedule-list.component.html
