+++ delectable-client/src/app/delectable/schedule/schedule.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { HttpClient, HttpParams } from '@angular/common/http';
 import { Observable } from 'rxjs';
-import { Schedule } from '../model/schedule';
+import { Schedule } from './schedule';
 
 @Injectable()
 export class ScheduleService {
