+++ delectable-client/src/main.ts
@@ -1,8 +1,7 @@
 import { enableProdMode } from '@angular/core';
 import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
-
-import { AppModule } from './app/app.module';
 import { environment } from './environments/environment';
+import { AppModule } from './app/delectable/app.module';
 
 if (environment.production) {
   enableProdMode();
